+++ hibernate-core/src/test/java/org/hibernate/test/bytecode/enhancement/lazy/proxy/inlinedirtychecking/LoadAndUpdateEntitiesWithCollectionsTest.java
@@ -63,6 +63,9 @@ public class LoadAndUpdateEntitiesWithCollectionsTest extends BaseNonConfigCoreF
 
 	@Before
 	public void setUp() {
+		if ( skipTest ) {
+			return;
+		}
 		inTransaction(
 				session -> {
 					User user = new User();
@@ -90,7 +93,10 @@ public class LoadAndUpdateEntitiesWithCollectionsTest extends BaseNonConfigCoreF
 	}
 
 	@After
-	public void tearDwon() {
+	public void tearDown() {
+		if ( skipTest ) {
+			return;
+		}
 		inTransaction(
 				session -> {
 					session.createQuery( "delete from SamplingOrder" ).executeUpdate();
@@ -103,6 +109,9 @@ public class LoadAndUpdateEntitiesWithCollectionsTest extends BaseNonConfigCoreF
 
 	@Test
 	public void testLoad() {
+		if ( skipTest ) {
+			return;
+		}
 		inTransaction(
 				session -> {
 					CriteriaBuilder cb = session.getCriteriaBuilder();
@@ -126,6 +135,9 @@ public class LoadAndUpdateEntitiesWithCollectionsTest extends BaseNonConfigCoreF
 
 	@Test
 	public void testAddUserRoles() {
+		if ( skipTest ) {
+			return;
+		}
 		inTransaction(
 				session -> {
 					SamplingOrder samplingOrder = getSamplingOrder( session );
@@ -200,6 +212,9 @@ public class LoadAndUpdateEntitiesWithCollectionsTest extends BaseNonConfigCoreF
 
 	@Test
 	public void testDeleteUserRoles() {
+		if ( skipTest ) {
+			return;
+		}
 		inTransaction(
 				session -> {
 					SamplingOrder samplingOrder = getSamplingOrder( session );
@@ -220,6 +235,9 @@ public class LoadAndUpdateEntitiesWithCollectionsTest extends BaseNonConfigCoreF
 
 	@Test
 	public void testModifyUserMail() {
+		if ( skipTest ) {
+			return;
+		}
 		inTransaction(
 				session -> {
 					SamplingOrder samplingOrder = getSamplingOrder( session );
